// backend/src/market/interfaces/coin-gecko-market-item.interface.ts
export interface CoinGeckoMarketItem {
  id: string;
  symbol: string;
  name: string;
  image: string | null;
  current_price: number | null;
  market_cap: number | null;
  market_cap_rank: number | null;
  total_volume: number | null;
  high_24h: number | null;
  low_24h: number | null;
  price_change_24h: number | null;
  price_change_percentage_24h: number | null;
  last_updated: string | null;
}